// src/shared/localization/enums.ts
export enum MESSAGE {
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  SIGNUP_SUCCESS = "SIGNUP_SUCCESS",
  CODE_SEND_SUCCESS = "CODE_SEND_SUCCESS",
  ACCOUNT_VERIFIED_SUCCESS = "ACCOUNT_VERIFIED_SUCCESS",
  PHONE_VERIFIED_SUCCESS = "PHONE_VERIFIED_SUCCESS",
  REST_PASSWORD_SUCCESS = "REST_PASSWORD_SUCCESS",
  UPDATE_SUCCESS = "UPDATE_SUCCESS",
  SEND_SUCCESS = "SEND_SUCCESS",
  JOIN_COMPANY_SUCCESS = "JOIN_COMPANY_SUCCESS",
  JOIN_COMPANY_PENDING = "JOIN_COMPANY_PENDING",
}